blog-post-generator/requirements.txt
google-genai>=1.68.0
Pillow>=10.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
streamlit>=1.30.0
selenium>=4.15.0
python-docx>=1.0.0
PyMuPDF>=1.23.0
